refactor: useModal & Modal 컴포넌트 수정
