+++ src/components/Main.tsx
@@ -7,12 +7,15 @@ import Modal from './common/Modal/Modal';
 import RestaurantDetail from './RestaurantDetail/RestaurantDetail';
 import { filterAndSortRestaurantList, useRestaurantList } from '../hooks/useRestaurantList';
 import { useSelectRestaurant } from '../hooks/useSelectRestaurant';
+import { useModal } from '../hooks/useModal';
+import { Restaurant } from '../types';
 
 const restaurantList = filterAndSortRestaurantList(getRestaurantListData());
 
 function Main() {
   const { currentRestaurantList, updateCurrentRestaurantList } = useRestaurantList(restaurantList);
   const { selectedRestaurant, setSelectedRestaurant } = useSelectRestaurant();
+  const { isModalOpen, openModal, closeModal } = useModal();
 
   useEffect(() => {
     window.addEventListener('beforeunload', () => {
@@ -20,13 +23,19 @@ function Main() {
     });
   }, []);
 
+  const handleRestaurantItem = (restaurant: Restaurant) => {
+    setSelectedRestaurant(restaurant);
+    openModal();
+  };
+
   return (
     <main>
       <FilterSection onChange={updateCurrentRestaurantList} />
-      <RestaurantList restaurantList={currentRestaurantList} onItemClick={setSelectedRestaurant} />
-      {selectedRestaurant && (
-        <Modal onClose={setSelectedRestaurant}>
-          <RestaurantDetail restaurant={selectedRestaurant} />
+      <RestaurantList restaurantList={currentRestaurantList} onItemClick={handleRestaurantItem} />
+
+      {isModalOpen && (
+        <Modal close={closeModal}>
+          {selectedRestaurant && <RestaurantDetail restaurant={selectedRestaurant} />}
         </Modal>
       )}
     </main>
